1/29/13 game change: tweak overall layout of challenges 1/29/13 game change: put challenge title above equations, on left 1/29/13 game change: put equations on left for all challenge types

+++ trunk/simulations-java/simulations/line-graphing/src/edu/colorado/phet/linegraphing/linegame/model/GraphTheLine.java
@@ -32,7 +32,7 @@ public class GraphTheLine extends Challenge {
     public GraphTheLine( String description, Line answer, EquationForm equationForm, ManipulationMode manipulationMode, IntegerRange xRange, IntegerRange yRange ) {
         super( createTitle( Strings.GRAPH_THE_LINE, manipulationMode ), description,
                answer, equationForm, manipulationMode, xRange, yRange,
-               new Point2D.Double( 700, 300 ), // origin offset
+               new Point2D.Double( 735, 300 ), // origin offset
                new Vector2D( xRange.getMin() + ( 0.65 * xRange.getLength() ), yRange.getMin() - 1 ), // point tool location 1
                new Vector2D( xRange.getMin() + ( 0.95 * xRange.getLength() ), yRange.getMin() - 4 ) );  // point tool location 2
     }
